requirements.txt
aiohttp>=3.8.6
python-dotenv==1.0.1
charset-normalizer>=3.3.1
frozenlist>=1.4.0
multidict>=6.0.4
numpy
pandas>=1.5.3
python-dateutil>=2.8.2
pytz>=2023.3.post1
tqdm>=4.64.1
yarl>=1.9.2
transformers>=4.46.2
google-api-python-client>=2.108.0
accelerate>=0.24.1
sentencepiece
huggingface-hub>=0.11.0
sentence-transformers>=2.2.2
scikit-learn>=1.0.1
matplotlib>=3.5.1
accelerate
bitsandbytes>=0.42.0
peft
litellm